test(aux-client): keep only the two end-to-end 401 eviction invariants
